Progress.

+++ WarehouseSystem.API/Controllers/OrderController.cs
@@ -1,7 +1,10 @@
-﻿using System.Web.Http;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Web.Http;
+using System.Web.Http.Description;
 using WarehouseSystem.API.Models;
 using WarehouseSystem.DataAccess;
-using WarehouseSystem.Repository;
 using WarehouseSystem.Repository.Interfaces;
 
 namespace WarehouseSystem.API.Controllers
@@ -17,38 +20,91 @@ namespace WarehouseSystem.API.Controllers
             this.productRepository = productRepository;
         }
 
-        public OrderModel Get(long id)
+        public IEnumerable<OrderModel> GetOrders()
         {
-            return GetOrderModel(id);
+            var orders = orderRepository.GetOrders();
+            var orderModels = new List<OrderModel>();
+
+            foreach (var order in orders)
+            {
+                var orderModel = CreateModelWithProductModel(order);
+                orderModels.Add(orderModel);
+            }
+            return orderModels;
         }
 
-        public void Put(long id, [FromBody]OrderModel orderModel)
+        [ResponseType(typeof(OrderModel))]
+        public IHttpActionResult Get(long id)
         {
+            var order = GetOrderModel(id);
+
+            if (order == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(order);
+        }
+
+        [ResponseType(typeof(void))]
+        public IHttpActionResult Put(long id, [FromBody]OrderModel orderModel)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             Order order = orderRepository.GetOrder(id);
+
+            if (order == null)
+            {
+                return NotFound();
+            }
+
             UpdateOrder(order, orderModel);
             orderRepository.Save(order);
+
+            return StatusCode(HttpStatusCode.NoContent);
         }
 
-        public void Post([FromBody]OrderModel orderModel)
+        [ResponseType(typeof(Order))]
+        public IHttpActionResult Post([FromBody]OrderModel orderModel)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             Order order = new Order();
             UpdateOrder(order, orderModel);
             orderRepository.Save(order);
+
+            return Ok(order);
         }
 
-        public void Delete(long id)
+        [ResponseType(typeof(Order))]
+        public IHttpActionResult Delete(long id)
         {
             Order order = orderRepository.GetOrder(id);
+
+            if (order == null)
+            {
+                return NotFound();
+            }
+
             orderRepository.Delete(order);
+            return Ok(order);
         }
 
         private OrderModel GetOrderModel(long id)
         {
-            return CreateModel(orderRepository.GetOrder(id));
+            var order = orderRepository.GetOrder(id);
+            return order == null ? null : CreateModel(orderRepository.GetOrder(id));
         }
 
         private void UpdateOrder(Order order, OrderModel orderModel)
         {
+            order.Id = orderModel.Id;
             order.ProductModelId = orderModel.ProductModelId;
             order.StoreName = orderModel.StoreName;
             order.ProductQuantity = orderModel.ProductQuantity;
@@ -59,9 +115,32 @@ namespace WarehouseSystem.API.Controllers
         {
             return new OrderModel
             {
+                Id = order.Id,
                 StoreName = order.StoreName,
                 ProductModelId = order.ProductModelId,
                 ProductQuantity = order.ProductQuantity
+            };
+        }
+
+        private OrderModel CreateModelWithProductModel(Order order)
+        {
+            return new OrderModel
+            {
+                Id = order.Id,
+                StoreName = order.StoreName,
+                ProductModelId = order.ProductModelId,
+                ProductQuantity = order.ProductQuantity,
+                ProductModel = CreateProductModel(order.ProductModel)
+            };
+        }
+
+        private ProductManufacturerModel CreateProductModel(ProductModel productModel)
+        {
+            return new ProductManufacturerModel
+            {
+                Id = productModel.Id,
+                Manufacturer = productModel.Manufacturer.Tittle,
+                ModelNumber = productModel.ModelNumber
             };
         }
     }
